Se realizaron mini pruebas
