+++ src/Backend/Controller/Controller.java
@@ -1,5 +1,10 @@
 package Backend.Controller;
 
+import Backend.DB.DB;
+
 public class Controller {
+    public static void main(String[] args) {
+        DB db = new DB();   
+    }
     
 }
